Refactored and made pip installable.
